+++ src/cli/index.ts
@@ -9,9 +9,9 @@ import { applyTask } from '../ai/apply-task';
 import { getModelConfig, getModelNames } from '../ai/model-config';
 import { redoLastTask } from '../ai/redo-task';
 import { runAIAssistedTask } from '../ai/task-workflow';
+import { undoTaskChanges } from '../ai/undo-task-changes';
 import { processFiles } from '../core/file-processor';
 import { generateMarkdown } from '../core/markdown-generator';
-import { undoTaskChanges } from '../git/undo-task-changes';
 import { runInteractiveMode } from '../interactive/interactive-workflow';
 import { DEFAULT_CACHE_PATH, clearCache } from '../utils/cache-utils';
 import { handleEditorAndOutput } from '../utils/editor-utils';
