+++ tests/unit/apply-task.test.ts
@@ -2,8 +2,8 @@ import path from 'node:path';
 import fs from 'fs-extra';
 import simpleGit from 'simple-git';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import * as applyChanges from '../../src/ai/apply-changes';
 import { applyTask } from '../../src/ai/apply-task';
-import * as applyChanges from '../../src/git/apply-changes';
 import * as gitTools from '../../src/utils/git-tools';
 
 vi.mock('../../src/utils/git-tools');
